Added success and fail outputs via uart

+++ include/sequence.h
@@ -6,8 +6,8 @@
 typedef struct{
     uint32_t state;
     uint32_t start_state;
-    uint32_t sequence_length;
-    uint32_t sequence_index;
+    uint16_t sequence_length;
+    uint16_t sequence_index;
     uint32_t seed;
 }LFSR;
